requirements.txt
crewai>=0.28.0
tavily-python>=0.3.0
requests>=2.31.0
node9>=0.1.0
